2.a2a-agents-with-python-sdk/requirements.txt
a2a-sdk>=0.3.0
click>=8.2.1
httpx>=0.28.1
a2a-sdk[http-server]
langchain>=0.3.27
langchain-google-genai>=2.1.9
langchain-tavily>=0.2.11
langgraph>=0.6.3
python-dotenv>=1.1.1
rich>=14.1.0
uvicorn>=0.35.0 